Book page
NOTE: About Enum:BookType.0~8

Since Type is of enum BookType, the system automatically converts the numerical value stored in the database (i.e., "0") to its corresponding value (i.e., "Undefined"). This is causing problems in displaying the localized value in en.json.

Hence, there is a need to change the # in Enum:BookType.# into the values indicated in the BookType class.